Added a way to stop the microphone and animations
Made the start button able to switch between start and stop states.
Added a play/stop sign to the start button depending on its state.
Added a global variable IS_PLAYING to main.js that tells whether
the microphone and animations are playing or not.
Changed the setBars() method from BarPlot.js to remove previous
nodes from the plot before setting the bars.